+++ schedulerx3-20240624/alibabacloud_schedulerx320240624/models.py
@@ -606,6 +606,7 @@ class CreateJobRequest(TeaModel):
         parameters: str = None,
         priority: int = None,
         route_strategy: int = None,
+        script: str = None,
         start_time: int = None,
         status: int = None,
         time_expression: str = None,
@@ -621,7 +622,6 @@ class CreateJobRequest(TeaModel):
         self.cluster_id = cluster_id
         self.description = description
         self.executor_block_strategy = executor_block_strategy
-        # This parameter is required.
         self.job_handler = job_handler
         # This parameter is required.
         self.job_type = job_type
@@ -634,6 +634,7 @@ class CreateJobRequest(TeaModel):
         self.parameters = parameters
         self.priority = priority
         self.route_strategy = route_strategy
+        self.script = script
         self.start_time = start_time
         self.status = status
         self.time_expression = time_expression
@@ -691,6 +692,8 @@ class CreateJobRequest(TeaModel):
             result['Priority'] = self.priority
         if self.route_strategy is not None:
             result['RouteStrategy'] = self.route_strategy
+        if self.script is not None:
+            result['Script'] = self.script
         if self.start_time is not None:
             result['StartTime'] = self.start_time
         if self.status is not None:
@@ -743,6 +746,8 @@ class CreateJobRequest(TeaModel):
             self.priority = m.get('Priority')
         if m.get('RouteStrategy') is not None:
             self.route_strategy = m.get('RouteStrategy')
+        if m.get('Script') is not None:
+            self.script = m.get('Script')
         if m.get('StartTime') is not None:
             self.start_time = m.get('StartTime')
         if m.get('Status') is not None:
@@ -776,6 +781,7 @@ class CreateJobShrinkRequest(TeaModel):
         parameters: str = None,
         priority: int = None,
         route_strategy: int = None,
+        script: str = None,
         start_time: int = None,
         status: int = None,
         time_expression: str = None,
@@ -791,7 +797,6 @@ class CreateJobShrinkRequest(TeaModel):
         self.cluster_id = cluster_id
         self.description = description
         self.executor_block_strategy = executor_block_strategy
-        # This parameter is required.
         self.job_handler = job_handler
         # This parameter is required.
         self.job_type = job_type
@@ -804,6 +809,7 @@ class CreateJobShrinkRequest(TeaModel):
         self.parameters = parameters
         self.priority = priority
         self.route_strategy = route_strategy
+        self.script = script
         self.start_time = start_time
         self.status = status
         self.time_expression = time_expression
@@ -854,6 +860,8 @@ class CreateJobShrinkRequest(TeaModel):
             result['Priority'] = self.priority
         if self.route_strategy is not None:
             result['RouteStrategy'] = self.route_strategy
+        if self.script is not None:
+            result['Script'] = self.script
         if self.start_time is not None:
             result['StartTime'] = self.start_time
         if self.status is not None:
@@ -902,6 +910,8 @@ class CreateJobShrinkRequest(TeaModel):
             self.priority = m.get('Priority')
         if m.get('RouteStrategy') is not None:
             self.route_strategy = m.get('RouteStrategy')
+        if m.get('Script') is not None:
+            self.script = m.get('Script')
         if m.get('StartTime') is not None:
             self.start_time = m.get('StartTime')
         if m.get('Status') is not None:
@@ -2055,6 +2065,285 @@ class GetDesigateInfoResponse(TeaModel):
             self.status_code = m.get('statusCode')
         if m.get('body') is not None:
             temp_model = GetDesigateInfoResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
+class GetJobExecutionRequest(TeaModel):
+    def __init__(
+        self,
+        app_name: str = None,
+        cluster_id: str = None,
+        job_execution_id: str = None,
+        mse_session_id: str = None,
+    ):
+        # This parameter is required.
+        self.app_name = app_name
+        # This parameter is required.
+        self.cluster_id = cluster_id
+        # This parameter is required.
+        self.job_execution_id = job_execution_id
+        self.mse_session_id = mse_session_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.app_name is not None:
+            result['AppName'] = self.app_name
+        if self.cluster_id is not None:
+            result['ClusterId'] = self.cluster_id
+        if self.job_execution_id is not None:
+            result['JobExecutionId'] = self.job_execution_id
+        if self.mse_session_id is not None:
+            result['MseSessionId'] = self.mse_session_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AppName') is not None:
+            self.app_name = m.get('AppName')
+        if m.get('ClusterId') is not None:
+            self.cluster_id = m.get('ClusterId')
+        if m.get('JobExecutionId') is not None:
+            self.job_execution_id = m.get('JobExecutionId')
+        if m.get('MseSessionId') is not None:
+            self.mse_session_id = m.get('MseSessionId')
+        return self
+
+
+class GetJobExecutionResponseBodyData(TeaModel):
+    def __init__(
+        self,
+        app_name: str = None,
+        attempt: int = None,
+        data_time: str = None,
+        duration: int = None,
+        end_time: str = None,
+        executor: str = None,
+        job_execution_id: str = None,
+        job_id: int = None,
+        job_name: str = None,
+        job_type: str = None,
+        parameters: str = None,
+        result: str = None,
+        route_strategy: int = None,
+        schedule_time: str = None,
+        server_ip: str = None,
+        start_time: str = None,
+        status: int = None,
+        time_type: int = None,
+        trigger_type: int = None,
+    ):
+        self.app_name = app_name
+        self.attempt = attempt
+        self.data_time = data_time
+        self.duration = duration
+        self.end_time = end_time
+        self.executor = executor
+        self.job_execution_id = job_execution_id
+        self.job_id = job_id
+        self.job_name = job_name
+        self.job_type = job_type
+        self.parameters = parameters
+        self.result = result
+        self.route_strategy = route_strategy
+        self.schedule_time = schedule_time
+        self.server_ip = server_ip
+        self.start_time = start_time
+        self.status = status
+        self.time_type = time_type
+        self.trigger_type = trigger_type
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.app_name is not None:
+            result['AppName'] = self.app_name
+        if self.attempt is not None:
+            result['Attempt'] = self.attempt
+        if self.data_time is not None:
+            result['DataTime'] = self.data_time
+        if self.duration is not None:
+            result['Duration'] = self.duration
+        if self.end_time is not None:
+            result['EndTime'] = self.end_time
+        if self.executor is not None:
+            result['Executor'] = self.executor
+        if self.job_execution_id is not None:
+            result['JobExecutionId'] = self.job_execution_id
+        if self.job_id is not None:
+            result['JobId'] = self.job_id
+        if self.job_name is not None:
+            result['JobName'] = self.job_name
+        if self.job_type is not None:
+            result['JobType'] = self.job_type
+        if self.parameters is not None:
+            result['Parameters'] = self.parameters
+        if self.result is not None:
+            result['Result'] = self.result
+        if self.route_strategy is not None:
+            result['RouteStrategy'] = self.route_strategy
+        if self.schedule_time is not None:
+            result['ScheduleTime'] = self.schedule_time
+        if self.server_ip is not None:
+            result['ServerIp'] = self.server_ip
+        if self.start_time is not None:
+            result['StartTime'] = self.start_time
+        if self.status is not None:
+            result['Status'] = self.status
+        if self.time_type is not None:
+            result['TimeType'] = self.time_type
+        if self.trigger_type is not None:
+            result['TriggerType'] = self.trigger_type
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AppName') is not None:
+            self.app_name = m.get('AppName')
+        if m.get('Attempt') is not None:
+            self.attempt = m.get('Attempt')
+        if m.get('DataTime') is not None:
+            self.data_time = m.get('DataTime')
+        if m.get('Duration') is not None:
+            self.duration = m.get('Duration')
+        if m.get('EndTime') is not None:
+            self.end_time = m.get('EndTime')
+        if m.get('Executor') is not None:
+            self.executor = m.get('Executor')
+        if m.get('JobExecutionId') is not None:
+            self.job_execution_id = m.get('JobExecutionId')
+        if m.get('JobId') is not None:
+            self.job_id = m.get('JobId')
+        if m.get('JobName') is not None:
+            self.job_name = m.get('JobName')
+        if m.get('JobType') is not None:
+            self.job_type = m.get('JobType')
+        if m.get('Parameters') is not None:
+            self.parameters = m.get('Parameters')
+        if m.get('Result') is not None:
+            self.result = m.get('Result')
+        if m.get('RouteStrategy') is not None:
+            self.route_strategy = m.get('RouteStrategy')
+        if m.get('ScheduleTime') is not None:
+            self.schedule_time = m.get('ScheduleTime')
+        if m.get('ServerIp') is not None:
+            self.server_ip = m.get('ServerIp')
+        if m.get('StartTime') is not None:
+            self.start_time = m.get('StartTime')
+        if m.get('Status') is not None:
+            self.status = m.get('Status')
+        if m.get('TimeType') is not None:
+            self.time_type = m.get('TimeType')
+        if m.get('TriggerType') is not None:
+            self.trigger_type = m.get('TriggerType')
+        return self
+
+
+class GetJobExecutionResponseBody(TeaModel):
+    def __init__(
+        self,
+        code: int = None,
+        data: GetJobExecutionResponseBodyData = None,
+        message: str = None,
+        request_id: str = None,
+        success: bool = None,
+    ):
+        self.code = code
+        # -\
+        self.data = data
+        self.message = message
+        # Id of the request
+        self.request_id = request_id
+        self.success = success
+
+    def validate(self):
+        if self.data:
+            self.data.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.code is not None:
+            result['Code'] = self.code
+        if self.data is not None:
+            result['Data'] = self.data.to_map()
+        if self.message is not None:
+            result['Message'] = self.message
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+        if self.success is not None:
+            result['Success'] = self.success
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Code') is not None:
+            self.code = m.get('Code')
+        if m.get('Data') is not None:
+            temp_model = GetJobExecutionResponseBodyData()
+            self.data = temp_model.from_map(m['Data'])
+        if m.get('Message') is not None:
+            self.message = m.get('Message')
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+        if m.get('Success') is not None:
+            self.success = m.get('Success')
+        return self
+
+
+class GetJobExecutionResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: GetJobExecutionResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = GetJobExecutionResponseBody()
             self.body = temp_model.from_map(m['body'])
         return self
 
@@ -4980,6 +5269,7 @@ class ListJobsResponseBodyDataRecords(TeaModel):
         parameters: str = None,
         priority: int = None,
         route_strategy: int = None,
+        script: str = None,
         status: int = None,
         time_expression: str = None,
         time_type: int = None,
@@ -5011,6 +5301,7 @@ class ListJobsResponseBodyDataRecords(TeaModel):
         self.parameters = parameters
         self.priority = priority
         self.route_strategy = route_strategy
+        self.script = script
         self.status = status
         self.time_expression = time_expression
         self.time_type = time_type
@@ -5074,6 +5365,8 @@ class ListJobsResponseBodyDataRecords(TeaModel):
             result['Priority'] = self.priority
         if self.route_strategy is not None:
             result['RouteStrategy'] = self.route_strategy
+        if self.script is not None:
+            result['Script'] = self.script
         if self.status is not None:
             result['Status'] = self.status
         if self.time_expression is not None:
@@ -5138,6 +5431,8 @@ class ListJobsResponseBodyDataRecords(TeaModel):
             self.priority = m.get('Priority')
         if m.get('RouteStrategy') is not None:
             self.route_strategy = m.get('RouteStrategy')
+        if m.get('Script') is not None:
+            self.script = m.get('Script')
         if m.get('Status') is not None:
             self.status = m.get('Status')
         if m.get('TimeExpression') is not None:
@@ -7866,6 +8161,7 @@ class UpdateJobRequest(TeaModel):
         parameters: str = None,
         priority: int = None,
         route_strategy: int = None,
+        script: str = None,
         start_time: int = None,
         time_expression: str = None,
         time_type: int = None,
@@ -7891,6 +8187,7 @@ class UpdateJobRequest(TeaModel):
         self.parameters = parameters
         self.priority = priority
         self.route_strategy = route_strategy
+        self.script = script
         self.start_time = start_time
         self.time_expression = time_expression
         self.time_type = time_type
@@ -7946,6 +8243,8 @@ class UpdateJobRequest(TeaModel):
             result['Priority'] = self.priority
         if self.route_strategy is not None:
             result['RouteStrategy'] = self.route_strategy
+        if self.script is not None:
+            result['Script'] = self.script
         if self.start_time is not None:
             result['StartTime'] = self.start_time
         if self.time_expression is not None:
@@ -7996,6 +8295,8 @@ class UpdateJobRequest(TeaModel):
             self.priority = m.get('Priority')
         if m.get('RouteStrategy') is not None:
             self.route_strategy = m.get('RouteStrategy')
+        if m.get('Script') is not None:
+            self.script = m.get('Script')
         if m.get('StartTime') is not None:
             self.start_time = m.get('StartTime')
         if m.get('TimeExpression') is not None:
@@ -8027,6 +8328,7 @@ class UpdateJobShrinkRequest(TeaModel):
         parameters: str = None,
         priority: int = None,
         route_strategy: int = None,
+        script: str = None,
         start_time: int = None,
         time_expression: str = None,
         time_type: int = None,
@@ -8052,6 +8354,7 @@ class UpdateJobShrinkRequest(TeaModel):
         self.parameters = parameters
         self.priority = priority
         self.route_strategy = route_strategy
+        self.script = script
         self.start_time = start_time
         self.time_expression = time_expression
         self.time_type = time_type
@@ -8100,6 +8403,8 @@ class UpdateJobShrinkRequest(TeaModel):
             result['Priority'] = self.priority
         if self.route_strategy is not None:
             result['RouteStrategy'] = self.route_strategy
+        if self.script is not None:
+            result['Script'] = self.script
         if self.start_time is not None:
             result['StartTime'] = self.start_time
         if self.time_expression is not None:
@@ -8146,6 +8451,8 @@ class UpdateJobShrinkRequest(TeaModel):
             self.priority = m.get('Priority')
         if m.get('RouteStrategy') is not None:
             self.route_strategy = m.get('RouteStrategy')
+        if m.get('Script') is not None:
+            self.script = m.get('Script')
         if m.get('StartTime') is not None:
             self.start_time = m.get('StartTime')
         if m.get('TimeExpression') is not None: